refactor(#65): item 3 — models/ split: loaders → lib/toy/io/loaders/
- toy_smollm2_loader.rb + toy_gpt2_loader.rb move under lib/toy/io/loaders/
  (loaders are I/O, not models); require sweep across lib/, demos/,
  examples/, Makefile (53 refs), docs/reference/loader-api.md.
- Configs/model classes (toy_smollm2.rb, toy_gpt2.rb, toy_vit.rb) STAY in
  lib/toy/models/ — least-breakage: dozens of requirers, no naming win.
- Mat (lib/toy/models/transformer.rb) STAYS: it requires ../ffi/tinynn
  (FFNFFICache + USE_FFI_MATMUL seam), so it FAILS the no-FFI-requires
  precondition for lib/toy/core/. Documented; a Mat/core split needs a
  tinynn decoupling pass first.
- DEPRECATED marking: audited every model-era class; NONE are clearly
  dead (pure-Ruby TransformerLM = live teaching/training path via
  toy_trainer + demos/train.rb; GPT2LM feeds tep_demo/openai_api.rb +
  GGUFLoad.load_gpt2; cache/result classes all internal to forward).
  Nothing falsely marked.

Gates (a699cf9): verify-mirrors, gate-compute-surface{,-cuda},
train_gate x4 + train_cuda_gate x3 byte-exact. Serve (f6d5eef):
serve_gate + serve_events_gate PASS.
