+++ demos/cuda_byo_ptr_smoke.rb
@@ -15,7 +15,7 @@
 # which modules are pulled in).
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy_smollm2_ffi_kv_cuda"
 require_relative "../lib/toy/train/training"
