+++ demos/qwen25_bench_cpu.rb
@@ -6,7 +6,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy_smollm2_ffi_kv"
 
 GGUF      = ENV["GGUF"]       || "data/qwen25-0.5b-native.gguf"
