+++ demos/qwen25_native_mmap.rb
@@ -11,7 +11,7 @@
 # Defaults to 0.5B F32 for quick smoke checks.
 
 require_relative "../lib/toy_smollm2_ffi_kv"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 
 GGUF  = ENV["GGUF"]  || "data/qwen25-0.5b-native.gguf"
 MAX_T = (ENV["MAX_T"] || "256").to_i
