+++ demos/qwen25_native_mmap_cuda.rb
@@ -12,7 +12,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy_smollm2_ffi_kv_cuda"
 
 GGUF  = ENV["GGUF"]  || "data/qwen25-1.5b-native.gguf"
