+++ demos/smollm2.rb
@@ -8,7 +8,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy/train/training"   # parse_ids
 
 GGUF     = "data/smollm2-135m-f32.gguf"
