+++ demos/smollm2_kv.rb
@@ -9,7 +9,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy_smollm2_ffi_kv"
 require_relative "../lib/toy/train/training"   # parse_ids
