+++ demos/smollm2_lora_train_bench_cuda.rb
@@ -16,7 +16,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy_smollm2_ffi_kv_cuda"
 
 GGUF      = ENV["GGUF"]      || "data/smollm2-135m-native.gguf"
