+++ demos/smollm2_lora_train_ce.rb
@@ -33,7 +33,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy_smollm2_ffi_kv"
 
 GGUF      = ENV["GGUF"]      || "data/smollm2-135m-native.gguf"
