+++ demos/smollm2_seq_parity.rb
@@ -22,7 +22,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy_smollm2_ffi_kv"
 require_relative "../lib/toy/llm/engine/llama_seq_engine"
