+++ demos/tinyllama.rb
@@ -3,7 +3,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy/train/training"
 
 GGUF     = "data/tinyllama-1.1b-f32.gguf"
