+++ examples/legacy/03_finetune_lora.rb
@@ -22,7 +22,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy/llm/engine/llama_seq_engine"
 
 GGUF      = ENV["GGUF"]    || "data/smollm2-135m-native.gguf"
