+++ examples/smoke_full_finetune.rb
@@ -14,7 +14,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy/llm/engine/llama_seq_engine"
 require_relative "../lib/toy/llm/adamw"
