+++ examples/smoke_recipe_lora.rb
@@ -26,7 +26,7 @@
 
 require_relative "../lib/toy"
 require_relative "../lib/toy/models/toy_smollm2"
-require_relative "../lib/toy/models/toy_smollm2_loader"
+require_relative "../lib/toy/io/loaders/toy_smollm2_loader"
 require_relative "../lib/toy/llm/engine/llama_seq_engine"
 require_relative "../lib/toy/llm/adamw"
 require_relative "../lib/toy/llm/recipes/lora"
