+++ lib/toy/compute_metal.rb
@@ -30,7 +30,7 @@ require_relative "../toy"
 
 # Models + configs (shared).
 require_relative "models/toy_smollm2"
-require_relative "models/toy_smollm2_loader"
+require_relative "io/loaders/toy_smollm2_loader"
 require_relative "models/toy_vit"
 
 # Engines — the Metal mirrors (generated by prep/gen_cuda_mirror.rb).
