+++ lib/toy/models/transformer_lm_cuda.rb
@@ -26,7 +26,7 @@ require_relative "arch"
 require_relative "../train/sampler"
 require_relative "../../toy"
 require_relative "toy_smollm2"
-require_relative "toy_smollm2_loader"
+require_relative "../io/loaders/toy_smollm2_loader"
 require_relative "../../toy_smollm2_ffi_kv_cuda"
 
 class ToyLMCuda
